backend/requirements.txt
celery==5.3.4
redis==5.0.0
attrs==23.2.0
azure-ai-ml==1.13.0
azure-ai-translation-text @ https://files.pythonhosted.org/packages/42/ed/6752793aba2150f3841dcad6b1a4ecfda22f7610314bdb1f078c48338aa3/azure_ai_translation_text-1.0.0b1-py3-none-any.whl
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-mgmt-authorization==4.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.15.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
idna==3.6
isodate==0.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
msal==1.27.0
msal-extensions==1.1.0
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydash==7.0.5
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
strictyaml==1.7.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
